Error in Lagrange Splitting is corrected
Interpolation with lagrange polynoms did not work our.
On one hand the values, that could not be interpolated were not set to
zero and stayed and stayed with a random number in the resampled signal.

Besides there was a nummeric error in Matlab withthe high nubers. Therefor
it was necessary to determine the polynoms differently.

r Please enter the commit message for your changes. Lines starti